% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi-internal.R
\name{bls.cov}
\alias{bls.cov}
\title{bls.cov
Performs backtracking line search along a pre-specified search direction}
\usage{
bls.cov(y0, x, mu, alpha, theta, mu.tmp, alpha.tmp, theta.tmp, b = 0.5,
  lambda1, lambda2, var.type, thresh = 1e-05, sc, nlevel, vt2)
}
\arguments{
\item{y0}{nxp observations matrix}

\item{x}{(np)xN matrix of covariates}

\item{mu}{real number, direction of update for offset}

\item{alpha}{direction of update for vector of regression parameters of length N}

\item{theta}{nxp matrix direction of update for matrix of interactions}

\item{mu.tmp}{real number, current offset}

\item{alpha.tmp}{length N vector, current regression parameters}

\item{theta.tmp}{nxp matrix, current matrix of interactions}

\item{b}{positive number in (0,1) factor by which the step size is reduced}

\item{lambda1}{positive number, regularization parameter for nuclear norm penalty}

\item{lambda2}{positive number, regularization parameter for l1 norm penalty}

\item{var.type}{vector of length p indicating column types for y (gaussian, binary, poisson)}

\item{thresh}{positive number, convergence criterion}

\item{sc}{scaling matrix}

\item{nlevel}{vector of integers indicating the number of levels of each factor in y}

\item{vt2}{vector indicating types of the columns of the extended data frame y (with dummies for every category)}
}
\value{
A list with the following elements
\item{mu}{the offset}
\item{alpha}{a (nb groups) x (nb variables) matrix containing the group effects}
\item{theta a}{(nb individuals) x (nb variables) matrix containing the individual effects}
\item{objective}{a vector containing the value of the objective function at every iteration}
\item{t}{the step size}
}
\description{
bls.cov
Performs backtracking line search along a pre-specified search direction
}
