% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi-internal.R
\name{bls.multi}
\alias{bls.multi}
\title{bls.multi
Performs backtracking line search along a pre-specified search direction}
\usage{
bls.multi(y0, groups, mu, alpha, theta, mu.tmp, alpha.tmp, theta.tmp,
  lambda1, lambda2, var.type, thresh = 1e-05, b = 0.5, nlevel, vt2)
}
\arguments{
\item{y0}{nxp matrix of observations}

\item{groups}{length n indicator factor of group memberships}

\item{mu}{real number new offset}

\item{alpha}{vector of length p*N(nb of groups) new regression parameter}

\item{theta}{nxp matrix new interaction matrix}

\item{mu.tmp}{real number current offset}

\item{alpha.tmp}{vector of length p*N(nb of groups) current regression parameter}

\item{theta.tmp}{nxp matrix current interaction matrix}

\item{lambda1}{positive number regularization parameter for nuclear norm penalty}

\item{lambda2}{positive number regularization parameter for l1 norm penalty}

\item{var.type}{vector of length p indicating variable types (gaussian, binary, poisson)}

\item{thresh}{positive number congervence criterion}

\item{b}{number in (0,1) factor by which step size is reduced}

\item{nlevel}{vector of integers indicating the number of levels of each factor in y}

\item{vt2}{vector indicating types of the columns of the extended data frame y (with dummies for every category)}
}
\description{
bls.multi
Performs backtracking line search along a pre-specified search direction
}
