% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practical_functions.R
\name{covmat}
\alias{covmat}
\title{covmat}
\usage{
covmat(XR = NULL, XC = NULL)
}
\arguments{
\item{XR}{(nxqr) matrix of row covariates}

\item{XC}{(pxqc) matrix of column covariates}
}
\value{
X, a ((np)x(qr+qc)) matrix combining XR and XC
}
\description{
covmat
}
\examples{
XR <- matrix(rnorm(10*3), nrow = 10)
XC <- matrix(rnorm(15*2), nrow = 15)
X <- covmat(XR, XC)
}
