% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi-tools.R
\name{covmatC}
\alias{covmatC}
\title{covmatC}
\usage{
covmatC(C, n)
}
\arguments{
\item{C}{nxK2 matrix of column covariates}

\item{n}{number of rows}
}
\value{
repeats C n times
}
\description{
covmatC
}
\examples{
C <- matrix(rnorm(10), 5)
cov <- covmatC(C,3)
}
