% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{covmatR}
\alias{covmatR}
\title{covmatR}
\usage{
covmatR(R, p)
}
\arguments{
\item{R}{nxK1 matrix of row covariates}

\item{p}{number ofcolumns}
}
\value{
repeats every row of R p times
}
\description{
covmatR
}
\examples{
R <- matrix(rnorm(10), 5)
cov <- covmatR(R,3)
}
