% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.mimi.cov}
\alias{cv.mimi.cov}
\title{cv.cov}
\usage{
cv.mimi.cov(y, x, nb.boot = 5, prob = 0.1, thresh = 1e-04, maxit = 100,
  var.type, lambda1.max = NULL, lambda2.max = NULL, lambda1.min = NULL,
  lambda2.min = NULL, size.grid = 10, length = 20, mu0 = NULL,
  alpha0 = NULL, theta0 = NULL, trace.it = F, offset = T, scale = F,
  max.rank = 20)
}
\arguments{
\item{y}{nxp observation matrix}

\item{x}{(np)xq matrix of covariates}

\item{nb.boot}{integer indicating the number of bootstrap samples for cross-validation}

\item{prob}{number in (0,1) proportion of missing entries added in the bootstrap samples}

\item{thresh}{positive number, convergence criterion}

\item{maxit}{integer, maximum number of iterations}

\item{var.type}{vector of length p indicating the types of the columns of y (gaussian, binary, poisson)}

\item{lambda1.max}{positive number, max value of nuclear norm regularization parameter}

\item{lambda2.max}{positive number, max value of l1 norm regularization parameter}

\item{lambda1.min}{positive number, min value of nuclear norm regularization parameter}

\item{lambda2.min}{positive number, min value of l1 norm regularization parameter}

\item{size.grid}{integer, length of the grid of lambda1 and lambda2 for cross-validation}

\item{length}{integer, length of the grid of lambda1 and lambda2}

\item{mu0}{real number, optional starting point for offset}

\item{alpha0}{Nxp matrix, optional starting point for group effects}

\item{theta0}{nxp matrix, optional starting point for interaction matrix}

\item{trace.it}{boolean, whether convergence information should be printed}

\item{offset}{boolean, whether an offset should be fitted (default FALSE)}

\item{scale}{boolean, whether column loss functions should be scaled (default FALSE)}

\item{max.rank}{integer, maximum rank of interaction matrix theta}
}
\value{
A list containing the following elements
\item{lambda1}{The selected value of lambda1 (regularization parameter for nuclear norm)}
\item{lambda2}{The selected value of lambda2 (regularization parameter for l1 norm)}
\item{estim.cv}{An object resulting from the mimi method, estimated with the selected values of lambda1 and lambda2}
\item{lambda1.grid}{A vector of size length containing the grid of lambda1 used for cv}
\item{lambda2.grid}{A vector of size length containing the grid of lambda2 used for cv}
}
\description{
cv.cov
}
\examples{
\dontrun{
n = 6; p = 2
y1 <- matrix(rnorm(mean = 0, n * p), nrow = n)
y2 <- matrix(rnorm(mean = 0, n * p), nrow = n)
y3 <- matrix(rnorm(mean = 2, n * p), nrow = n)
y <- cbind(matrix(rnorm(mean = c(y1), n * p), nrow = n),
           matrix(rbinom(n * p, prob = c(exp(y2)/(1+exp(y2))), size = 1), nrow = n),
           matrix(rpois(n * p, lambda = c(exp(y3))), nrow = n))
var.type <- c(rep("gaussian", p), rep("binary", p), rep("poisson", p))
idx_NA <- sample(1:(3 * n * p), size = round(0.1 * 3 * n * p))
y[idx_NA] <- NA
x <- matrix(rnorm(6*6*2), nrow = 6*6)
res <- cv.mimi.cov(y, x, var.type = var.type, maxit = 2, nb.boot = 2, thresh =1)
}
}
