% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.mimi.multi}
\alias{cv.mimi.multi}
\title{cv.mimi.multi}
\usage{
cv.mimi.multi(y, groups, nb.boot = 10, prob = 0.1, thresh = 1e-04,
  maxit = 100, var.type, lambda1.max = NULL, lambda2.max = NULL,
  lambda1.min = NULL, lambda2.min = NULL, size.grid = 20,
  length = 20, mu0 = NULL, alpha0 = NULL, theta0 = NULL,
  trace.it = F, offset = F, scale = F, max.rank = 20)
}
\arguments{
\item{y}{nxp observations matrix}

\item{groups}{factor of length n indicating groups}

\item{nb.boot}{integer indicating the number of bootstrap samples for cross-validation}

\item{prob}{number in (0,1) proportion of missing entries added in the bootstrap samples}

\item{thresh}{positive number, convergence criterion}

\item{maxit}{integer, maximum number of iterations}

\item{var.type}{vector of length p indicating the types of the columns of y (gaussian, binary, poisson)}

\item{lambda1.max}{positive number, max value of nuclear norm regularization parameter}

\item{lambda2.max}{positive number, max value of l1 norm regularization parameter}

\item{lambda1.min}{positive number, min value of nuclear norm regularization parameter}

\item{lambda2.min}{positive number, min value of l1 norm regularization parameter}

\item{size.grid}{integer, length of the grid of lambda1 and lambda2 for cross-validation}

\item{length}{integer, length of the grid of lambda1 and lambda2 for warm start}

\item{mu0}{real number, optional starting point for offset}

\item{alpha0}{Nxp matrix, optional starting point for group effects}

\item{theta0}{nxp matrix, optional starting point for interaction matrix}

\item{trace.it}{boolean, whether convergence information should be printed}

\item{offset}{boolean, whether an offset should be fitted (default FALSE)}

\item{scale}{boolean, whether column loss functions should be scaled (default FALSE)}

\item{max.rank}{integer, maximum rank of interaction matrix theta}
}
\value{
A list containing the following elements
\item{lambda1}{The selected value of lambda1 (regularization parameter for nuclear norm)}
\item{lambda2}{The selected value of lambda2 (regularization parameter for l1 norm)}
\item{estim.cv}{An object resulting from the mimi method, estimated with the selected values of lambda1 and lambda2}
\item{lambda1.grid}{A vector of size length containing the grid of lambda1 used for cv}
\item{lambda2.grid}{A vector of size length containing the grid of lambda2 used for cv}
}
\description{
cv.mimi.multi
}
\examples{
\dontrun{
y <- matrix(rnorm(4 * 3), nrow = 4)
y[sample(1:12, size = 3)] <- NA
groups <- c(1,1,2,2)
groups <- as.factor(groups)
vt <- rep("gaussian", 3)
res <- cv.mimi.multi(y, groups, 2, var.type = vt, maxit = 1, size.grid = 2, thresh = 1)
}
}
