% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{Frob}
\alias{Frob}
\title{Frob
compute stopping criterion of weighted softImpute}
\usage{
Frob(Uold, Dsqold, Vold, U, Dsq, V)
}
\arguments{
\item{Uold}{nxq matrix of left singular vectors}

\item{Dsqold}{vector singular values (q)}

\item{Vold}{pxq matrix of right singular vectors}

\item{U}{nxq matrix of left singular vectors}

\item{Dsq}{vector of singular values (q)}

\item{V}{pxq matrix of right singular vectors}
}
\value{
value of stopping criterion
}
\description{
Frob
compute stopping criterion of weighted softImpute
}
\examples{
U <- matrix(rnorm(5*2), nrow = 5)
V <- matrix(rnorm(3*2), nrow = 3)
Dsq <- abs(rnorm(2))
Uold <- matrix(rnorm(5*2), nrow = 5)
Vold <- matrix(rnorm(3*2), nrow = 3)
Dsqold <- abs(rnorm(2))
crit <- Frob(Uold,Dsqold,Vold,U,Dsq,V)
}
