% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi-internal.R
\name{irwls.multi}
\alias{irwls.multi}
\title{irwls.multi}
\usage{
irwls.multi(y, groups, var.type, lambda1, lambda2, nlevel = NULL,
  maxit = 100, mu0 = NULL, alpha0 = NULL, theta0 = NULL,
  thresh = 1e-05, trace.it = F, offset = F, scale = F,
  max.rank = 5, vt2 = NULL)
}
\arguments{
\item{y}{observation matrix}

\item{groups}{factor indicating group memberships, if vector will be treated as a factor}

\item{var.type}{vector of size p indicating types of columns in y (gaussian, binary, poisson)}

\item{lambda1}{positive number, value of the nuclear norm regularization parameter}

\item{lambda2}{positive number, value of the l1 norm regularization parameter}

\item{nlevel}{vector of integers indicating the number of levels of each factor in y}

\item{maxit}{integer, maximum number of iterations}

\item{mu0}{real number, initial value of the offset, default 0}

\item{alpha0}{matrix of size (nb of groups)x(number of variables), initial value of the group effect, default 0}

\item{theta0}{matrix of size (nb of ind.)x(number of variables), initial value of the individual effect, default 0}

\item{thresh}{positive number, convergence criterion}

\item{trace.it}{boolean, if TRUE information about convergence will be displayes, default FALSE}

\item{offset}{boolean, whether an offset should be fitted}

\item{scale}{boolean indicating whether or not the column loss functions should be scaled}

\item{max.rank}{integer, maximum rank of interaction matrix theta}

\item{vt2}{vector indicating types of the columns of the extended data frame y (with dummies for every category)}
}
\value{
A list vith the following elements
\item{y.imputed}{the original data matrix where missing entries are imputed by their estimated means}
\item{param}{the estimated parameter matrix}
\item{mu}{the offset}
\item{alpha}{a (nb groups) x (nb variables) matrix containing the group effects}
\item{theta}{a (nb individuals) x (nb variables) matrix containing the individual effects}
}
\description{
irwls.multi
}
