% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{log_factorial}
\alias{log_factorial}
\title{log_factorial}
\usage{
log_factorial(x)
}
\arguments{
\item{x}{matrix of integers}
}
\value{
the log of the factorial of every entry in x
}
\description{
log_factorial
}
\examples{
x <- matrix(seq(1, 10), nrow = 5)
res <- log_factorial(x)
}
