% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi.R
\name{mimi.lr}
\alias{mimi.lr}
\title{mimi.lr
Compute solution of mimi for low-rank model along regularization path}
\usage{
mimi.lr(y, var.type = c("gaussian", "binary", "categorical", "poisson"),
  lambda1, maxit = 100, theta0 = NULL, thresh = 1e-05, trace.it = F,
  lambda1.max = NULL, length = 20, upper = 12, lower = -12,
  offset = F, scale = F, max.rank = 20, wmax = NULL)
}
\arguments{
\item{y}{nxp observation matrix}

\item{var.type}{vector of length p indicating types of y columns (gaussian, binary, poisson)}

\item{lambda1}{positive number, regularization parameter for nuclear norm penalty}

\item{maxit}{integer, maximum number of iterations}

\item{theta0}{matrix of size nxp, initial interactions (optional)}

\item{thresh}{positive number, convergence criterion}

\item{trace.it}{boolean, whether convergence information should be printed}

\item{lambda1.max}{positive number, max value of regularization parameter for nuclear norm penalty}

\item{length}{integer, size of grid for regularization path}

\item{upper}{real number, upper bound on parameters entries}

\item{lower}{real number, lower bound on parameters entries}

\item{offset}{boolean, whether or not an offset should be fitted, default FALSE}

\item{scale}{boolean indicating whether or not the column loss functions should be scaled}

\item{max.rank}{integer, maximum rank of interaction matrix}
}
\value{
A list with the following elements
\item{lambda1.grid}{the grid of lambda1 (log scale)}
\item{list.res}{List of same length as regularization grid with output of mimi for each lambda}
\item{list.theta}{List of same length as regularization grid with value of theta for each lambda}
}
\description{
mimi.lr
Compute solution of mimi for low-rank model along regularization path
}
\examples{
y0 <- matrix(rnorm(6 * 10), nrow = 6)
y0[sample(1:50, size = 10)] <- NA
var.type <- rep("gaussian", 10)
res <- mimi.lr(y0, var.type, lambda1 = 0.1)
}
