% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi-internal.R
\name{quad_approx}
\alias{quad_approx}
\title{quad_approx}
\usage{
quad_approx(y, param, var.type)
}
\arguments{
\item{y}{observation matrix (nxP) - extended with dummies for cat variables}

\item{param}{matrix of parameters (nxP) - extended for cat variables}

\item{var.type}{type of the variables in y (length p<=P) - NOT extended for cat variables}
}
\value{
matrix of weights for the quadratic approximation (nxP)
}
\description{
quad_approx
}
