% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwls_l1_nuc.cov.R
\name{rwls_l1_nuc.cov}
\alias{rwls_l1_nuc.cov}
\title{rwls_l1_nuc.cov}
\usage{
rwls_l1_nuc.cov(y, x, var.type, lambda1, lambda2, nlevel = NULL,
  maxit = 100, mu0 = NULL, alpha0 = NULL, theta0 = NULL,
  thresh = 1e-05, trace.it = F, offset = F, scale = F,
  max.rank = 20, vt2 = NULL, wmax = NULL)
}
\arguments{
\item{y}{nxp observation matrix}

\item{x}{(np)xN covariates matrix}

\item{var.type}{vector of size p indicating types of columns in y (gaussian, binary, poisson)}

\item{lambda1}{positive number, value of the nuclear norm regularization parameter}

\item{lambda2}{positive number, value of the l1 norm regularization parameter}

\item{nlevel}{vector of integers indicating the number of levels of each factor in y}

\item{maxit}{integer, maximum number of iterations}

\item{mu0}{real number, initial value of the offset, default 0}

\item{alpha0}{matrix of size (nb of groups)x(number of variables), initial value of the group effect, default 0}

\item{theta0}{matrix of size (nb of ind.)x(number of variables), initial value of the individual effect, default 0}

\item{thresh}{positive number, convergence criterion}

\item{trace.it}{boolean, if TRUE information about convergence will be displayes, default FALSE}

\item{offset}{boolean, if TRUE offset is computed, otherwise set to 0, default FALSE}

\item{scale}{boolean indicating whether or not the column loss functions should be scaled}

\item{max.rank}{integer, maximum rank of interaction matrix theta}

\item{vt2}{vector indicating types of the columns of the extended data frame y (with dummies for every category)}
}
\value{
A list vith the following elements
\item{y}{the original data matrix}
\item{y.imputed}{the original data matrix where missing entries are imputed by their estimated means}
\item{param}{the estimated parameter matrix}
\item{mu}{the offset}
\item{alpha}{a (nb groups) x (nb variables) matrix containing the group effects}
\item{theta}{a (nb individuals) x (nb variables) matrix containing the individual effects}
\item{objective}{a vector containing the value of the objective function at every iteration}
}
\description{
rwls_l1_nuc.cov
}
\examples{
n = 6; p = 2
y1 <- matrix(rnorm(mean = 0, n * p), nrow = n)
y2 <- matrix(rnorm(mean = 0, n * p), nrow = n)
y3 <- matrix(rnorm(mean = 2, n * p), nrow = n)
y <- cbind(matrix(rnorm(mean = c(y1), n * p), nrow = n),
           matrix(rbinom(n * p, prob = c(exp(y2)/(1+exp(y2))), size = 1), nrow = n),
           matrix(rpois(n * p, lambda = c(exp(y3))), nrow = n))
var.type <- c(rep("gaussian", p), rep("binary", p), rep("poisson", p))
idx_NA <- sample(1:(3 * n * p), size = round(0.1 * 3 * n * p))
y[idx_NA] <- NA
x <- matrix(rnorm(6*6*2), nrow = 6*6)
nl <- rep(1, 6)
res <- rwls_l1_nuc.cov(y, x, var.type, 0.1, 0.2, nl)
}
