% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{wlra}
\alias{wlra}
\title{wlra}
\usage{
wlra(x, w = NULL, lambda = 0, x0 = NULL, thresh = 1e-05, maxit = 100,
  rank.max = NULL)
}
\arguments{
\item{x}{nxp matrix to be approximates}

\item{w}{nxp matrix of weights (optional)}

\item{lambda}{regularization parameter}

\item{x0}{nxp matrix of initial values}

\item{thresh}{convergence criterion}

\item{maxit}{maximum number of iterations}

\item{rank.max}{maximum number of sv to compute}
}
\value{
a list containing the following elements
\item{d}{the vector of singular values}
\item{u}{the left singular vectors}
\item{v}{the right singular vectors}
\item{convergence}{boolean indicating if algorithm converged before maxit iterations}
\item{iter}{number of iterations performed}
}
\description{
wlra
}
\examples{
x <- matrix(rnorm(10*6), nrow = 10)
x[sample(60, 10)] <- NA
res <- wlra(x)
}
