% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwls_l1_nuc.cov.R
\name{wls_l1_nuc.cov}
\alias{wls_l1_nuc.cov}
\title{wls_l1_nuc.cov}
\usage{
wls_l1_nuc.cov(y, x, lambda1, lambda2, weights = NULL, thresh = 1e-05,
  maxit = 100, mu0 = NULL, alpha0 = NULL, theta0 = NULL,
  trace.it = F, offset = F, max.rank = 20)
}
\arguments{
\item{y}{observation matrix}

\item{x}{covariates matrix}

\item{lambda1}{positive number, value of the nuclear norm regularization parameter}

\item{lambda2}{positive number, value of the l1 norm regularization parameter}

\item{weights}{matrix of size (nb of ind.)x(number of variables) with entries in (0,1), optional weights matrix}

\item{thresh}{positive number, convergence criterion}

\item{maxit}{integer, maximum number of iterations}

\item{mu0}{real number, initial value of the offset, default 0}

\item{alpha0}{a vector of length (nc covariates): initial regression parameter, default 0}

\item{theta0}{matrix of size (nb of ind.)x(number of variables), initial value of the individual effect, default 0}

\item{trace.it}{boolean, if TRUE information about convergence will be displayes, default FALSE}

\item{offset}{boolean, if TRUE offset is computed, otherwise set to 0, default FALSE}

\item{max.rank}{integer, maximum rank of interaction matrix}
}
\value{
A list vith the following elements
\item{y}{the original data matrix}
\item{y.imputed}{the original data matrix where missing entries are imputed by their estimated means}
\item{param}{the estimated parameter matrix}
\item{mu}{the offset}
\item{alpha}{a vector of length (nc covariates): the regression parameter}
\item{theta}{a (nb individuals) x (nb variables) matrix containing the individual effects}
\item{objective}{a vector containing the value of the objective function at every iteration}
}
\description{
wls_l1_nuc.cov
}
\examples{
y <- matrix(rnorm(6 * 10), nrow = 6)
y[sample(1:50, size = 10)] <- NA
x <- matrix(rnorm(60*2), nrow = 60)
res <- wls_l1_nuc.cov(y, x, lambda1 = 0.1, lambda2 = 0.2)
}
