% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi-internal.R
\name{armijo.lr}
\alias{armijo.lr}
\title{armijo.lr
Performs Armijo backtracking line search along a pre-specified search direction}
\usage{
armijo.lr(y0, theta, theta.tmp, w, z, b = 0.5, lambda1, var.type,
  thresh = 1e-05, nlevel, vt2)
}
\arguments{
\item{y0}{nxp observations matrix}

\item{theta}{nxp matrix direction of update for matrix of interactions}

\item{theta.tmp}{nxp matrix, current matrix of interactions}

\item{b}{positive number in (0,1) factor by which the step size is reduced}

\item{lambda1}{positive number, regularization parameter for nuclear norm penalty}

\item{var.type}{vector of length p indicating column types for y (gaussian, binary, poisson)}

\item{thresh}{positive number, convergence criterion}

\item{nlevel}{vector of integers indicating the number of levels of each factor in y}

\item{vt2}{vector indicating types of the columns of the extended data frame y (with dummies for every category)}
}
\value{
A list with the following elements
\item{theta}{(nb individuals) x (nb variables) matrix containing the individual effects}
\item{objective}{a vector containing the value of the objective function at every iteration}
\item{t}{the step size}
}
\description{
armijo.lr
Performs Armijo backtracking line search along a pre-specified search direction
}
