% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{bls.multi}
\alias{bls.multi}
\title{bls.multi
Performs backtracking line search along a pre-specified search direction}
\usage{
bls.multi(y0, groups, mu, alpha, theta, mu.tmp, alpha.tmp, theta.tmp, lambda1,
  lambda2, var.type, thresh = 0.001, b = 0.5)
}
\arguments{
\item{y0}{nxp matrix of observations}

\item{groups}{length n indicator factor of group memberships}

\item{mu}{real number new offset}

\item{alpha}{vector of length p*N(nb of groups) new regression parameter}

\item{theta}{nxp matrix new interaction matrix}

\item{mu.tmp}{real number current offset}

\item{alpha.tmp}{vector of length p*N(nb of groups) current regression parameter}

\item{theta.tmp}{nxp matrix current interaction matrix}

\item{lambda1}{positive number regularization parameter for nuclear norm penalty}

\item{lambda2}{positive number regularization parameter for l1 norm penalty}

\item{var.type}{vector of length p indicating variable types (gaussian, binary, poisson)}

\item{thresh}{positive number congervence criterion}

\item{b}{number in (0,1) factor by which step size is reduced}
}
\value{
A list with the following elements
\item{mu}{the offset}
\item{alpha}{a (nb groups) x (nb variables) matrix containing the group effects}
\item{theta a}{(nb individuals) x (nb variables) matrix containing the individual effects}
\item{objective}{a vector containing the value of the objective function at every iteration}
\item{t}{the step size}
}
\description{
bls.multi
Performs backtracking line search along a pre-specified search direction
}
\examples{
y0 <- matrix(rnorm(6 * 10), nrow = 6)
m <- 0
a <- matrix(rnorm(30), nrow = 3)
t <- matrix(rnorm(6 * 10), nrow = 6)
mt <- 0.1
at <- matrix(rnorm(30), nrow = 3)
tt <- matrix(rnorm(6 * 10), nrow = 6)
v <- rep("gaussian", 10)
gps <- as.factor(c(1,1,2,2,3,3))
bls <- bls.multi(y0, gps, m, a, t, mt, at, tt, 1, 1, v)
}
