% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi-tools.R
\name{covmat}
\alias{covmat}
\title{covmat}
\usage{
covmat(R = NULL, C = NULL, n, p)
}
\arguments{
\item{R}{nxK1 matrix of row covariates}

\item{C}{nxK2 matrix of column covariates}

\item{n}{number of rows}

\item{p}{number ofcolumns}
}
\value{
the joint product of R and C, a (np)x(K1+K1) matrix in order row1col1,row2col1,...,rowncol1, row1col2, row2col2,...,rowncolp
}
\description{
covmat
}
\examples{
R <- matrix(rnorm(10), 5)
C <- matrix(rnorm(9), 3)
covs <- covmat(R,C,5,3)
}
