% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practical_functions.R, R/side_functions.R
\name{covmat}
\alias{covmat}
\title{covmat}
\usage{
covmat(R, C, n, p)

covmat(R, C, n, p)
}
\arguments{
\item{R}{nxK1 matrix of row covariates}

\item{C}{nxK2 matrix of column covariates}

\item{n}{number of rows}

\item{p}{number ofcolumns}

\item{XR}{(nxqr) matrix of row covariates}

\item{XC}{(pxqc) matrix of column covariates}
}
\value{
X, a ((np)x(qr+qc)) matrix combining XR and XC

the joint product of R and C, a (np)x(K1+K1) matrix in order row1col1,row2col1,...,rowncol1, row1col2, row2col2,...,rowncolp
}
\description{
covmat

covmat
}
\examples{
XR <- matrix(rnorm(10*3), nrow = 10)
XC <- matrix(rnorm(15*2), nrow = 15)
X <- covmat(XR, XC)
R <- matrix(rnorm(10), 5)
C <- matrix(rnorm(9), 3)
covs <- covmat(R,C,5,3)
}
