% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv.mimi}
\alias{cv.mimi}
\title{cv.mimi
Perform cross-validation for mimi model with covariates/groups along regularization path}
\usage{
cv.mimi(y, model = c("groups", "covariates"), x = NULL, groups = NULL,
  var.type, lambda1.min = NULL, lambda2.min = NULL, lambda1.max = NULL,
  lambda2.max = NULL, maxit = 100, mu0 = NULL, alpha0 = NULL,
  theta0 = NULL, thresh = 1e-04, trace.it = F, length = 20,
  offset = F, nb.boot = 10, prob = 0.1, size.grid = 20)
}
\arguments{
\item{y}{nxp matrix of observations}

\item{model}{either one of "groups" or "covariates", indicating which model should be fitted}

\item{x}{(np)xN matrix of covariates (optional)}

\item{groups}{factor of length n indicating groups}

\item{var.type}{vector of length p indicating the types of the columns of y (gaussian, binomial, poisson)}

\item{lambda1.min}{positive number, min value of nuclear norm regularization parameter}

\item{lambda2.min}{positive number, min value of l1 norm regularization parameter}

\item{lambda1.max}{positive number, max value of nuclear norm regularization parameter}

\item{lambda2.max}{positive number, max value of l1 norm regularization parameter}

\item{maxit}{integer, maximum number of iterations}

\item{mu0}{real number, optional starting point for offset}

\item{alpha0}{Nxp matrix, optional starting point for group effects}

\item{theta0}{nxp matrix, optional starting point for interaction matrix}

\item{thresh}{positive number, convergence criterion}

\item{trace.it}{boolean, whether convergence information should be printed}

\item{length}{integer, length of the grid of lambda1 and lambda2 for warm start}

\item{offset}{boolean, whether an offset should be fitted}

\item{nb.boot}{integer indicating the number of bootstrap samples for cross-validation}

\item{prob}{number in (0,1) proportion of missing entries added in the bootstrap samples}

\item{size.grid}{integer, length of the grid of lambda1 and lambda2 for cross-validation}
}
\value{
A list with the following elements
\item{lambda1.grid}{the grid of lambda1 used for warm start (log scale)}
\item{lambda2.grid}{the grid of lambda2 used for warm start (log scale)}
\item{list.res}{list of results of rwls_l1_nuc (same length as grid of lambda)}
\item{list.mu}{list of offsets (same length as grid of lambda)}
\item{list.alpha}{list of regression parameters (same length as grid of lambda)}
\item{list.theta}{list of interaction matrices (same length as grid of lambda)}
}
\description{
cv.mimi
Perform cross-validation for mimi model with covariates/groups along regularization path
}
\examples{
n = 6; p = 2
y1 <- matrix(rnorm(mean = 0, n * p), nrow = n)
y2 <- matrix(rnorm(mean = 0, n * p), nrow = n)
y3 <- matrix(rnorm(mean = 2, n * p), nrow = n)
y <- cbind(matrix(rnorm(mean = c(y1), n * p), nrow = n),
           matrix(rbinom(n * p, prob = c(exp(y2)/(1+exp(y2))), size = 1), nrow = n),
           matrix(rpois(n * p, lambda = c(exp(y3))), nrow = n))
var.type <- c(rep("gaussian", p), rep("binomial", p), rep("poisson", p))
idx_NA <- sample(1:(3 * n * p), size = round(0.1 * 3 * n * p))
y[idx_NA] <- NA
x <- matrix(rnorm(6*6*2), nrow = 6*6)
res <- cv.mimi(y, model = "cov", x = x, var.type = var.type, maxit = 2, nb.boot = 2, size.grid = 2)
}
