% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi-internal.R
\name{irwls.lr}
\alias{irwls.lr}
\title{irwls.lr}
\usage{
irwls.lr(y, var.type, lambda1, nlevel = NULL, maxit = 100,
  theta0 = NULL, thresh = 1e-05, trace.it = F, offset = F,
  scale = F, max.rank = 20, vt2 = NULL)
}
\arguments{
\item{y}{nxp observation matrix}

\item{var.type}{vector of size p indicating types of columns in y (gaussian, binary, poisson)}

\item{lambda1}{positive number, value of the nuclear norm regularization parameter}

\item{nlevel}{vector of integers indicating the number of levels of each factor in y}

\item{maxit}{integer, maximum number of iterations}

\item{theta0}{matrix of size (nb of ind.)x(number of variables), initial value of the individual effect, default 0}

\item{thresh}{positive number, convergence criterion}

\item{trace.it}{boolean, if TRUE information about convergence will be displayes, default FALSE}

\item{offset}{boolean, if TRUE offset is computed, otherwise set to 0, default FALSE}

\item{scale}{boolean indicating whether or not the column loss functions should be scaled}

\item{max.rank}{integer, maximum rank of interaction matrix theta}

\item{vt2}{vector indicating types of the columns of the extended data frame y (with dummies for every category)}
}
\value{
A list vith the following elements
\item{y.imputed}{the imputed data set}
\item{theta}{estimated parameter matrix}
}
\description{
irwls.lr
}
