% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mimi.R
\name{mimi}
\alias{mimi}
\title{mimi
Compute solution of mimi with covariates/groups along regularization path}
\usage{
mimi(y, model = c("groups", "covariates", "low-rank"), x = NULL,
  groups = NULL, var.type = c("gaussian", "binary", "categorical",
  "poisson"), lambda1, lambda2, maxit = 100, mu0 = NULL, alpha0 = NULL,
  theta0 = NULL, thresh = 1e-04, trace.it = F, lambda1.max = NULL,
  lambda2.max = NULL, length = 20, upper = 12, lower = -12,
  offset = F, scale = F, max.rank = 20, wmax = NULL)
}
\arguments{
\item{y}{nxp matrix of observations}

\item{model}{either one of "groups" or "covariates", indicating which model should be fitted}

\item{x}{(np)xN matrix of covariates (optional)}

\item{groups}{factor of length n indicating groups (optional)}

\item{var.type}{vector of length p indicating the data types of the columns of y (gaussian, binary or poisson)}

\item{lambda1}{positive number regularization parameter for nuclear norm penalty}

\item{lambda2}{positive number regularization parameter for l1 norm penalty}

\item{maxit}{integer maximum number of iterations}

\item{mu0}{real number initial value of offset (optional)}

\item{alpha0}{vector of length N: initial value of regression parameter (optional)}

\item{theta0}{matrix of size nxp: initial value of interactions (optional)}

\item{thresh}{positive number, convergence criterion}

\item{trace.it}{boolean indicating whether convergence information should be printed}

\item{lambda1.max}{positive number starting value of regularization parameter}

\item{lambda2.max}{positive number starting value of regularization parameter}

\item{length}{length of the grid of regularization parameters}

\item{upper}{real number upper bound on entries of alpha and theta}

\item{lower}{real number lower bound on entries of alpha and theta}

\item{offset}{boolean indicating whether or not an offset should be fitted}

\item{scale}{boolean indicating whether or not the column loss functions should be scaled}

\item{max.rank}{integer, maximum rank of interaction matrix theta}
}
\value{
A list with the following elements
\item{lambda1.grid}{the grid of lambda1 used for warm start (log scale)}
\item{lambda2.grid}{the grid of lambda2 used for warm start (log scale)}
\item{list.res}{list of results of rwls_l1_nuc (same length as grid of lambda)}
\item{list.mu}{list of offsets (same length as grid of lambda)}
\item{list.alpha}{list of regression parameters (same length as grid of lambda)}
\item{list.theta}{list of interaction matrices (same length as grid of lambda)}
}
\description{
mimi
Compute solution of mimi with covariates/groups along regularization path
}
\examples{
n = 6; p = 2
y1 <- matrix(rnorm(mean = 0, n * p), nrow = n)
y2 <- matrix(rnorm(mean = 0, n * p), nrow = n)
y3 <- matrix(rnorm(mean = 2, n * p), nrow = n)
y <- cbind(matrix(rnorm(mean = c(y1), n * p), nrow = n),
           matrix(rbinom(n * p, prob = c(exp(y2)/(1+exp(y2))), size = 1), nrow = n),
           matrix(rpois(n * p, lambda = c(exp(y3))), nrow = n))
var.type <- c(rep("gaussian", p), rep("binary", p), rep("poisson", p))
idx_NA <- sample(1:(3 * n * p), size = round(0.1 * 3 * n * p))
y[idx_NA] <- NA
x <- matrix(rnorm(6*6*2), nrow = 6*6)
res <- mimi(y, model = "covariates", x = x, var.type = var.type, lambda1 = 0.1, lambda2 = 0.2)
}
