% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{quad_approx}
\alias{quad_approx}
\title{quad_approx}
\usage{
quad_approx(y, param, var.type)
}
\arguments{
\item{y}{column of observation (length n)}

\item{param}{column of parameters (length n)}

\item{var.type}{type of the variables in y (length 1)}
}
\value{
vector of weights for the quadratic approximation
}
\description{
quad_approx
}
\examples{
y <- rnorm(6)
param <- rnorm(6)
var.type <- "gaussian"
wghts <- quad_approx(y, param, var.type)
}
