% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/side_functions.R
\name{wght}
\alias{wght}
\title{wght}
\usage{
wght(y, param, var.type)
}
\arguments{
\item{y}{real number: observation}

\item{param}{real number: current value of parameter}

\item{var.type}{type of variable y (gaussian, binomial, poisson)}
}
\value{
weight of the quadratic approximation
}
\description{
wght
}
\examples{
y <- rnorm(1)
param <- 0.5
var.type = "gaussian"
wght <- wght(y, param, var.type)
}
